+++ GDX.Tests/EditMode/CoreTests.cs
@@ -2,6 +2,7 @@
 // dotBunny licenses this file to you under the BSL-1.0 license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using NUnit.Framework;
 
 namespace GDX
@@ -15,6 +16,31 @@ namespace GDX
         [Category(Core.TestCategory)]
         public void InitializeOnMainThread_AlreadyInitialized_EarlyOut()
         {
+            Type core = Reflection.GetType("GDX.Core");
+            Assert.IsNotNull(core);
+
+            Reflection.SetFieldValue(null, core, "s_InitializedMainThread", true,
+                Reflection.PrivateStaticFlags);
+            Assert.IsTrue(Reflection.GetFieldValue<bool>(null, core, "s_InitializedMainThread",
+                Reflection.PrivateStaticFlags));
+
+
+            Core.InitializeOnMainThread();
+        }
+
+        [Test]
+        [Category(Core.TestCategory)]
+        public void InitializeOnMainThread_NotAlreadyInitialized_Executes()
+        {
+            Type core = Reflection.GetType("GDX.Core");
+            Assert.IsNotNull(core);
+
+            Reflection.SetFieldValue(null, core, "s_InitializedMainThread", false,
+                Reflection.PrivateStaticFlags);
+
+            Assert.IsFalse(Reflection.GetFieldValue<bool>(null, core, "s_InitializedMainThread",
+                Reflection.PrivateStaticFlags));
+
             Core.InitializeOnMainThread();
         }
     }