+++ GDX/Reflection.cs
@@ -18,9 +18,9 @@ namespace GDX
         /// </summary>
         public const BindingFlags PrivateFieldFlags = BindingFlags.Instance | BindingFlags.NonPublic;
         /// <summary>
-        ///     <see cref="BindingFlags"/> for an internal static.
+        ///     <see cref="BindingFlags"/> for a private static.
         /// </summary>
-        public const BindingFlags InternalStaticFlags = BindingFlags.Static | BindingFlags.NonPublic;
+        public const BindingFlags PrivateStaticFlags = BindingFlags.Static | BindingFlags.NonPublic;
         /// <summary>
         ///     <see cref="BindingFlags"/> for a public static.
         /// </summary>
@@ -44,27 +44,20 @@ namespace GDX
         ///     Access the field value of a specific <paramref name="targetObject"/>, which may not be normally accessible.
         /// </summary>
         /// <remarks></remarks>
-        /// <param name="targetObject">The instanced object which will have it's field value read.</param>
-        /// <param name="type">The explicit type of the <paramref name="targetObject"/>.</param>
+        /// <param name="targetObject">The instanced object which will have it's field value read; use a null value if this is a static field.</param>
+        /// <param name="type">The qualified type of the <paramref name="targetObject"/>.</param>
         /// <param name="name">The field's name to read.</param>
         /// <param name="flags">The field's access flags.</param>
         /// <typeparam name="T">The type of data being read from the field.</typeparam>
         /// <returns>The field's value.</returns>
-        public static T GetFieldValue<T>(object targetObject, string type, string name, BindingFlags flags = PrivateFieldFlags)
+        public static T GetFieldValue<T>(object targetObject, Type type, string name, BindingFlags flags = PrivateFieldFlags)
         {
-            Assembly[] loadedAssemblies = AppDomain.CurrentDomain.GetAssemblies();
-            int loadedAssembliesCount = loadedAssemblies.Length;
-            for (int i = 0; i < loadedAssembliesCount; i++)
+            if (type == null)
             {
-                Type targetType = loadedAssemblies[i].GetType(type);
-                if (targetType == null)
-                {
-                    continue;
-                }
-                FieldInfo field = targetType.GetField(name, flags);
-                return (T)field?.GetValue(targetObject);
+                return default;
             }
-            return default;
+            FieldInfo field = type.GetField(name, flags);
+            return (T)field?.GetValue(targetObject);
         }
 
         /// <summary>
@@ -82,6 +75,7 @@ namespace GDX
         {
             if (targetObject == null)
                 return null;
+
             Type type = targetObject.GetType();
             FieldInfo f = type.GetField(name, fieldFlags);
             if (f != null)
@@ -91,6 +85,26 @@ namespace GDX
 
             PropertyInfo p = type.GetProperty(name,propertyFlags);
             return p == null ? null : p.GetValue(targetObject, null);
+        }
+
+        /// <summary>
+        ///     Access the field value of a specific <paramref name="targetObject"/>, which may not be normally accessible.
+        /// </summary>
+        /// <remarks></remarks>
+        /// <param name="targetObject">The instanced object which will have it's field value read; use a null value if this is a static property.</param>
+        /// <param name="type">The explicit type of the <paramref name="targetObject"/>.</param>
+        /// <param name="name">The field's name to read.</param>
+        /// <param name="flags">The field's access flags.</param>
+        /// <typeparam name="T">The type of data being read from the field.</typeparam>
+        /// <returns>The field's value.</returns>
+        public static T GetPropertyValue<T>(object targetObject, Type type, string name, BindingFlags flags = PrivateFieldFlags)
+        {
+            if (type == null)
+            {
+                return default;
+            }
+            PropertyInfo propertyInfo = type.GetProperty(name, flags);
+            return (T)propertyInfo?.GetValue(targetObject);
         }
 
         /// <summary>
@@ -124,24 +138,98 @@ namespace GDX
         public static object InvokeStaticMethod(string type, string method, object[] parameters = null,
             BindingFlags flags = PublicStaticFlags)
         {
-            Assembly[] loadedAssemblies = AppDomain.CurrentDomain.GetAssemblies();
-            int loadedAssembliesCount = loadedAssemblies.Length;
-            for (int i = 0; i < loadedAssembliesCount; i++)
+            Type targetType = GetType(type);
+            if (targetType != null)
             {
-                Type targetType = loadedAssemblies[i].GetType(type);
-
-                // For code coverage purposes were not going to invert this!
-                // ReSharper disable once InvertIf
-                if (targetType != null)
+                MethodInfo targetMethod = targetType.GetMethod(method, flags);
+                if (targetMethod != null)
                 {
-                    MethodInfo targetMethod = targetType.GetMethod(method, flags);
-                    if (targetMethod != null)
-                    {
-                        return targetMethod.Invoke(null, parameters ?? Core.EmptyObjectArray);
-                    }
+                    return targetMethod.Invoke(null, parameters ?? Core.EmptyObjectArray);
                 }
             }
             return null;
+        }
+
+        /// <summary>
+        ///     Set the field or property value of a specific <paramref name="targetObject"/>, which may not be
+        ///     normally accessible.
+        /// </summary>
+        /// <param name="targetObject">The instanced object which will have it's field or property value set.</param>
+        /// <param name="name">The field or property's name to set.</param>
+        /// <param name="value">The value to set the field or property to.</param>
+        /// <param name="fieldFlags">The field's access flags.</param>
+        /// <param name="propertyFlags">The property's access flags.</param>
+        /// <returns>true/false if the value was set.</returns>
+        public static bool SetFieldOrPropertyValue(object targetObject, string name, object value,
+            BindingFlags fieldFlags = PrivateFieldFlags, BindingFlags propertyFlags = PrivateFieldFlags)
+        {
+            if (targetObject == null)
+                return false;
+
+            Type type = targetObject.GetType();
+            FieldInfo f = type.GetField(name, fieldFlags);
+            if (f != null)
+            {
+                f.SetValue(targetObject, value);
+                return true;
+            }
+
+            PropertyInfo p = type.GetProperty(name,propertyFlags);
+            if (p != null)
+            {
+                p.SetValue(targetObject, value);
+                return true;
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        ///     Set the field value of a specific <paramref name="targetObject"/>, which may not be normally accessible.
+        /// </summary>
+        /// <param name="targetObject">The instanced object which will have it's field value set; use a null value if this is a static field.</param>
+        /// <param name="type">The explicit type of the <paramref name="targetObject"/>.</param>
+        /// <param name="name">The field's name to set.</param>
+        /// <param name="value">The value to set the field to.</param>
+        /// <param name="flags">The field's access flags.</param>
+        /// <returns>true/false if the value was set.</returns>
+        public static bool SetFieldValue(object targetObject, Type type, string name, object value,
+            BindingFlags flags = PrivateFieldFlags)
+        {
+            if (type != null)
+            {
+                FieldInfo field = type.GetField(name, flags);
+                if (field != null)
+                {
+                    field.SetValue(targetObject, value);
+                    return true;
+                }
+            }
+            return false;
+        }
+
+        /// <summary>
+        ///     Set the field value of a specific <paramref name="targetObject"/>, which may not be normally accessible.
+        /// </summary>
+        /// <param name="targetObject">The instanced object which will have it's field value set; use a null value if this is a static property.</param>
+        /// <param name="type">The type of the <paramref name="targetObject"/>.</param>
+        /// <param name="name">The field's name to set.</param>
+        /// <param name="value">The value to set the field to.</param>
+        /// <param name="flags">The field's access flags.</param>
+        /// <returns>true/false if the value was set.</returns>
+        public static bool SetPropertyValue(object targetObject, Type type, string name, object value,
+            BindingFlags flags = PrivateFieldFlags)
+        {
+            if (type != null)
+            {
+                PropertyInfo property = type.GetProperty(name, flags);
+                if (property != null)
+                {
+                    property.SetValue(targetObject, value);
+                    return true;
+                }
+            }
+            return false;
         }
     }
 }